+++ app/models/__init__.py
@@ -2,4 +2,4 @@ from .db import db
 from .user import User
 from .project import Project
 from .comment import Comment
-from .updates import Update
+from .update import Update
